requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.30.1
py-solc-x==1.1.1
web3==6.18.0
